requirements.txt
fastapi==0.118.0
onnxruntime==1.23.0
optimum[onnxruntime]
pillow==11.0.0
streamlit==1.50.0
transformers==4.53.3
uvicorn==0.37.0
python-multipart